Add GetHuman{Size,Speed}AndSuffix()
This is a complement to GetHuman{Size,Speed}Representation except instead of
returning a string it returns the value (as float64) and suffix (as
string).

This allows the calling code to format the float64 portion in a custom
way instead of the hardcoded "%.2f".

Example:

fmt.Println(x.GetHumanSizeRepresentation()) => "63.24 GiB"

size, suffix := x.GetHumanSizeAndSuffix()
fmt.Printf("%0.1f %s", size, suffix) => "63.2 GiB"